requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-CORS==5.0.0
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
